Extract DPM wave client routes
